requirements.txt
Flask==2.2.3
paho-mqtt==2.2.1
Werkzeug==2.2.3
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2 